Add option to use HQPRES for Training Center screen
The Community Highlander now has a hook in `HQPRES.ShowPromotionUI()
that allows mods to override the screen used for promotions and the
Training Center. This change allows mods to force Squad Select to use
that function when going to the Training Center so the correct screen is
displayed (they may override the base game hero promotion screen with
New Promotion Screen By Default for example).

Resolves #10.